+++ src/i18n/MControlCenter_en.ts
@@ -239,5 +239,9 @@
         <source>Choose the mode that best suits your use case</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Cooling</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 </TS>