+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 - Added fan mode display
 - Added description on mode tab
+- Added "Cooler Boost" item to tray menu
 - Fixed display of the window when clicking the tray icon if the application was minimized
 
 ## [0.3.1] 2022-11-04